requirements.txt
beautifulsoup4==4.9.3
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
decorator==4.4.2
gTTS==2.1.1
gTTS-token==1.1.3
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
moviepy==1.0.3
numpy==1.19.2
Pillow==8.0.1
praw==7.1.0
prawcore==1.5.0
proglog==0.1.9
requests==2.24.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
tqdm==4.50.2
update-checker==0.18.0
urllib3==1.25.11
websocket-client==0.57.0
